requirements.txt
aiofiles==0.4.0
cachetools==3.1.1
certifi==2019.9.11
chardet==3.0.4
h11==0.8.1
h2==3.1.1
hpack==3.0.0
httpcore==0.3.0
httptools==0.0.13
hyperframe==5.2.0
idna==2.8
motor==2.0.0
multidict==4.6.1
pymongo==3.9.0
python-dotenv==0.10.3
requests==2.22.0
requests-async==0.5.0
rfc3986==1.3.2
sanic==19.9.0
tornado==6.0.3
ujson==1.35
urllib3==1.25.7
uvloop==0.14.0
websockets==8.1
